src/main/java/org/nikiforova/solutions/easy/linkedList/MergeTwoSortedLists.java
package org.nikiforova.solutions.easy.linkedList;

public class MergeTwoSortedLists {
    public static ListNode mergeTwoLists(ListNode list1, ListNode list2) {
        ListNode head = new ListNode();
        ListNode currentNode = head;
        while (list1 != null && list2 != null) {
            if (list1.val <= list2.val) {
                currentNode.next = list1;
                list1 = list1.next;
            } else {
                currentNode.next = list2;
                list2 = list2.next;
            }
            currentNode = currentNode.next;
        }
        currentNode.next = list1 == null ? list2 : list1;
        return head.next;
    }
}
